// packages/core/src/errorboundary/ArchbaseErrorBoundaryContext.d.ts
export type ArchbaseErrorBoundaryContextType = {
    didCatch: boolean;
    error: any;
    resetErrorBoundary: (...args: any[]) => void;
};
export declare const ArchbaseErrorBoundaryContext: import("react").Context<ArchbaseErrorBoundaryContextType>;
//# sourceMappingURL=ArchbaseErrorBoundaryContext.d.ts.map